backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic[email]==2.10.3
pydantic-settings==2.7.0
faker==33.1.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
alembic==1.14.0
slowapi==0.1.9
celery==5.4.0
openpyxl==3.1.5
redis==5.2.1
boto3==1.35.99
flower==2.0.1
# tortoise-orm==0.21.7  # alternative async ORM — swap in via container.py get_connection()
# asyncpg==0.30.0       # required driver for tortoise-orm with Postgres
pytest==8.3.4
pytest-cov==6.0.0
httpx==0.28.1
ruff==0.9.1
mypy==1.14.1
